app/requirements.txt
Django==4.2.13
psycopg2-binary==2.9.9
Pillow==10.4.0
pytesseract==0.3.10
boto3==1.34.0
django-storages==1.14.2
gunicorn==21.2.0
